+++ NEWS.md
@@ -3,6 +3,8 @@
 ## Bug fixes
 
 * Repair `subtarget_hashes.cross()` for crosses on a single grouping variable.
+* Repair dynamic `group()` used with specialized formats (#1236, @adamaltmejd).
+
 
 # Version 7.12.0